+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - On Wayland, Add HiDPI cursor support
 - On Web, add the ability to query "Light" or "Dark" system theme send `ThemeChanged` on change.
 - Fix `Event::to_static` returning `None` for user events.
+- On Wayland, Hide CSD for fullscreen windows.
 
 # 0.21.0 (2020-02-04)